backend/requirements.txt
# ─── Core Framework ───────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ─── State Store ──────────────────────────────────────────────────
redis==5.1.1

# ─── Routing Solver ──────────────────────────────────────────────
ortools==9.10.4067

# ─── ML / Data Science ──────────────────────────────────────────
scikit-learn==1.5.2
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3

# ─── Real-Time Communication ────────────────────────────────────
python-socketio==5.11.4

# ─── HTTP Client (external API calls) ───────────────────────────
httpx==0.27.2

# ─── Diff Engine ─────────────────────────────────────────────────
deepdiff==8.0.1

# ─── Environment / Config ───────────────────────────────────────
python-dotenv==1.0.1
pydantic-settings==2.5.2
